# cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Login Gateway Service",
        "title": "Login Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9001",
    "basePath": "/",
    "paths": {
        "/mintegra/login": {
            "post": {
                "description": "Create Authenticate user, store in redis, and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Login With Redis",
                "parameters": [
                    {
                        "description": "Login Post Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.LoginMainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of User by filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master User"
                ],
                "summary": "Get Master User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer No",
                        "name": "customer_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Site",
                        "name": "site",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Department",
                        "name": "department",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Section",
                        "name": "section",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Position",
                        "name": "position",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Crew",
                        "name": "crew",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status Data",
                        "name": "status_data",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Source",
                        "name": "source",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "\"2006-01-02 15:04:05\"",
                        "description": "Last Update Time",
                        "name": "last_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.TUserMainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create one or more master User entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master User"
                ],
                "summary": "Create Master User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer No",
                        "name": "customer_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Post Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_request.TUserPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_royanqodri_Login-Gateway-API_model_request.LoginRequest": {
            "type": "object",
            "required": [
                "menu",
                "password",
                "username"
            ],
            "properties": {
                "menu": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_request.TUserPostDetailRequest": {
            "type": "object",
            "required": [
                "blood_type",
                "email_address",
                "id_customer",
                "name",
                "password",
                "phone_no",
                "type",
                "username"
            ],
            "properties": {
                "blood_type": {
                    "type": "string"
                },
                "email_address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_customer": {
                    "type": "integer"
                },
                "insert_by": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_no": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "update_by": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_request.TUserPostRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_request.TUserPostDetailRequest"
                    }
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_response.LoginMainResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.LoginResponse"
                    }
                },
                "data_header": {},
                "status_response": {
                    "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_response.LoginResponse": {
            "type": "object",
            "properties": {
                "customer_no": {
                    "type": "string"
                },
                "id_customer": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_response.MainResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "data_header": {},
                "status_response": {
                    "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "status_text": {
                    "type": "string"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_response.TUserGetResponse": {
            "type": "object",
            "properties": {
                "blood_type": {
                    "type": "string"
                },
                "customer_no": {
                    "type": "string"
                },
                "email_address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_customer": {
                    "type": "integer"
                },
                "insert_by": {
                    "type": "string"
                },
                "insert_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_no": {
                    "type": "string"
                },
                "status_data": {
                    "type": "string"
                },
                "update_by": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_royanqodri_Login-Gateway-API_model_response.TUserMainResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.TUserGetResponse"
                    }
                },
                "data_header": {},
                "status_response": {
                    "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}